Extend the README with basic usage and installation steps
Most importantly, emphasize on the fact that the plugin runs with
forked (modded) avocado(-vt) repos and thus needs extra patches
on these in order to run. This is the present situation at the
time of 67.0 and will hopefully change in the future once we
coordinate more with the original repositories.