Refactor primal test nodes from the runner and loader
We unify these in terms of naming, construction places, treating
methods, and boolean checks.

Using 0-prepended names, we free up keywords like the "scan", "root",
and "preinstall" for use of custom test nodes. The 0 is used to comply
with the numbering scheme and for the primal nodes only while the rest
of the internal nodes are either never directly parsed in configuration
or have their other peculiarities.