Use leave/nonleave test sets and noop manual step
In the past the "none" test restriction set was associated with
internal nodes but at the same time with running no (leaf) tests,
both with different internal implementations (similarly to the
different internal implementations we had for parsing and running
primal test nodes). This can be very confusing so let us make sure
we are explicit that "all" represents "all leaves", "none" is all
"nonleaves" (internal and original test nodes), and not running
tests (or doing anything else) is equivalent to a noop manual step.