Use "join" and "suffix" operators for heterogeneous Cartesian variants
These are operators native to the Cartesian config that were added
in 2015 in order to support heterogenous variants for a test suite
regarding the SPICE protocol. We use this to rely on more native
approaches for variant heterogenization than a plethora of extra
steps patching the final test parameters.

This usage of suffix simplifies the parameter definitions and the
usage of join deprecates most of their in-code customization. All
these extra functions are no longer needed with the native operators.