Rewrite the parameter parser module with reparsable configuration in mind
The new parameter parser module minimizes memory storage and allows
for reproducible creation of multiple identical parsers/params by
storing only the process of obtaining any given configuration. It
also unified the year-long patching of the params parser interface
into a simple object hierarchy focused on reparsable configuration.

This saves us from the old parser customization approaches which are
now rendered impossible by a degrading recursive dictionary generator
that `get_dicts` has turned into. The original method uses a dangerous
attribute mutation to differentiate the first call to this method from
recursive calls. One way to fix it and to enable multiple calls to the
method is by restoring the attribute.

However, another fallout happens even when recreating the generator if
we don't make sure to perform a deeper than usual parser copy. This in
turn has larger performance penalty, all in conclusion that no more
workarounds make sense and we can just rewrite the module with the new
paradigm of reparsable configuration which it was gradually patched
towards.