Replace all remaining parameter customization with optimized substitution
The Cartesian config has its own optimized "set" operation which
will automatically replace any pre-existing parameters in the
value of a newly set parameter. Let's take advantage of that.

We now use host aware parameters only in the configuration rather
than in code because we want to minimize the implicit in-code
customization (some of which very hard to trace) for the sake of
explicit in-configuration customization for each vm.