i2n: Replace the vmnet class with an intranator-extended version
This injects code from the inherited class similarly to the way
we were customizing the env processing before. The extended class
is located in and managed by the test suite. It is not the nicest
approach but it works and doesn't require us to modify a patch
each time we modify the extended class.