Remove setters for left, right, and general tunnel parameters
The parameter handling of the entire vmnet is already known to
require some useful unification but the essential parameters for
a tunnel remain the left and right node parameters. Thus, replacing
the left or right end point for the tunnel will also change its
configuration.

Making the configuration and parameter hangling easier in the future
remains in the roadmap for the entire vmnet subpackage and to some
extend also for the cartgraph subpackage.