Split the normal test set into GUI and non-GUI variants
The previous variant arrangment will run "all" tests if either "gui"
or "nongui" test set is selected on the command line while it should
in fact run "all normal" tests.

Such refinements of a normal run are also more up to the user with
the current two being just an example of a possible split.