Rename all bridges using a new enumeration-based nomenclature
According to this new nomenclature, we will use index names for bridges
as b1, b2, b3, etc. similarly to the way we did it for vms (vm1, vm2,
vm3, etc.) and this will allow for an arbitrary number of bridges with
a possible choice of name for each newly added bridge.