Drop default cdroms parameter to avoid transferring it to unwanted vms
These are used only for install tests when they can be overriden
and there can be collision when for instance a vm would define
configuration for some cd1 which can then propagate to other vms.
We are safer when we explicitly define the CD for a vm or in install
tests which make use of only one vm.

This has happened even before the Avocado 70.0 update but now let's
be explicit as we are already explicit about vm configuration.