Be explicit about all within-code nonleaf restrictions
This is not a perfect protection but it will significanty reduce the
chance of collisions if a user test suite makes use of some of the
internal node keywords.

It simplifies previously unclear handling of objectless restrctions
where in fact what was meant was nonleaf restrictions. These are
now explicitly passed at all dependent places in the code base.