Add a final badge to reflect our last plugin migration todo
We already knew the current unit test coverage is in a very
rudimentary state as all unit tests we have are the originally
migrated ones and this part hasn't been touched ever since. It
is clearly something in the TODO together with actually running
the sample test provider (test suite) at the CI stage.