Remove explicit config file installation in develop mode
We do what the other plugins do in the case of local (develop)
installation, namely provide relative path defaults that will
work well with the preinstalled package structure.

The handling is similar to the way we detect config files and
use the test suite path from the settings which if not provided
will be replaced with a relative path from the current module so
we reuse this test suite path preparation.