Use SKIP status for test dependencies with abort policy
The previously used TestAbortError is treated as regular TestError
which is not very useful when trying to distinguish tests that were
not executed (skipped) due to failed setup nodes. The reasoning now
is that SKIP only implies a test was skipped for some reason which
could be failed setup, expected bug, or something else that signals
that the test wasn't executed and the validating coverage is available.