Isolate initial state types to allow for exclusively on/off 0-states
Changing the configuration in the test sets config should now allow
for on-state and off-state only graphs in addition to the default
mixed graph.

The only extra difficulty was in the the fact that original install
tests must end with a shutdown requiring an extra clean (qq) boot
in the case of exclusively on 0-states.