Abort the entire run on missing permanent (manually provided) vms
The trick was how to abort the traversal considering that there are
phantom parts of the Autotest codebase responsible for abort and still
around in Avocado VT but dormant and not really used (our coupling
with them while still active is no longer functional). The way to do
it was to be a bit move clever about the shared root (scan) node - it
is the place where we scan for dependencies and decide which ones to
cancel. So? Well, if we have the data of the nodes and which ones are
canceled still around at the time of the reason to abort, we could just
go through all nodes once again and cancel all of them!