Refactor the default vms and tests variants and parameters
Once the command line choices are parsed, the missing but required
values are filled with appropriate defaults from the configs and
this refactoring makes this more obvious and better tested.

Thanks to the long awaited tool template refactoring in the intertest
setup we could finally reduce the diff lines for all tools. Too bad
this wasn't done earlier as it was already enabled with the explicitly
importable function list.