Provide complete default vm restrictions through the available vms
This is necessary for tests that need to parse graphs beyond the
current selection of vms, e.g. the update tool that also includes
a cleanup (reversed traversal) of additional nodes.

The information about selection and availability of vms is now
completely contained in the vm strings of selected and available
variants.