Complete the parameter dictionary explicitness during parsing
After all of this refactoring and (cmd parser originated) argument
standardization the final implicit steps of passing the parameter
dictionary implicitly as a string to test objects and nodes can now
be pinpointed and dealt with as the ultimate such step.

Even though there is a known problem with the implementation of
suffix operators, we can still isolate it better to the exact place
in code where it is becoming such.