Use leaf-neutral primary test restrictions to allow complete graph parsing
As not all setup tests leaving vm states for reuse are expected
to be internal nodes, there is one best way to arrange the primary
test restrictions and this involves for "all" and "none" to truly
be what they mean instead of a mere facade for leaf/nonleaf variants.

In particular, "none" is excluded to best represent empty Cartesian
products and "all" has now been fully enabled to access all tests
variants in the configuration. We can thus parse any type of node
(internal, shared root, etc.) from the command line without any
artificial blockades and any nodes that are not connected to the
shared root will not be run, all for a simpler, easier, and explicit
understanding of the Cartesian graph structure.