Add minimal multivariant vm configuration in the sample test suite
This is not meant as the real multivariant vm support which is still
limited and tracked by a TODO but about adding the posibility to choose
different vm default variants.

The TODO describes it all but with the joint operators we might now
enable multi-vm-variant tests, something that would require development
and improved concepts though.