Adapt to the new job and test suite APIs
We use this adaptation process to also drop redundant and no longer
effective job handling on our side. The more we leave to the default
backend, the better.

In addition, we now explicitly use a test suite and load the
tests participating in it, making sure the loading has the logdir
location provided to dump the parsed graph.