Add ability to retry/rerun tests
With two extra parameters it's possible to tell the runner to
retry a test N times.

The parameters are `retry_attempts` and `retry_stop`. The first is
the maximum number of retries, and the second indicates when to
stop retrying. The possible combinations of these values are:

- `retry_stop = error`: retry until error or a maximum of
                        `retry_attempts` number of times
- `retry_stop = success`: retry until success or a maximum of
                          `retry_attempts` number of times
- `retry_stop = none`: retry a maximum of `retry_attempts` number
                       of times

Only tests with the status of pass, warning, error or failure will
be retried. Other status will be ignored and the test will thus
run only once.