Allow for dynamic suite path in the test suite configs
This would significantly reduce the hardcoding of the test suite path
which is a major step towards an old wish of our test developers.

However, it can't be enabled in the configs for now (for which we
have a second commit) due to limitations of the current join operators
and a pending review of a PR on the Avocado VT side. One day we will
get rid of this heavy and sluggish dependency but it won't be today.