Reconsider boot states as actual on root states
The previous expectation that we could use only on states or only
off states is not valid as on states always make use of off state
(at the very least an off root state). This also significantly
simplifies separate handling of boot/root states based on fixed
names, redundant root get/set/unset types, as well as handling of
on backend root (boot) states within the state setup and not by
the backends themselves.