Implement separate (optional) pool modification operations
This requires tweaking the behavior of the regular state operations
like always making sure that the remote root state has been made
local too.

The local and pool root set/unset operations are mutually exclusive
and the all pool updates are triggered with a custom parameter with
the knowledge of the user.

We error out if we update the pool without any local root state.