Implement state pool operation locks with optional readonly mode
Since the pool is shared among containers or other parallel isolated
execution environments, it is essential to protect from undesired
IO concurrency and use this as a defaul behavior. If however, we
are using a readonly pool, waiting on locks can be traded for speed
at the responsibility of the user.