Drop the custom scan parsing and running for ad-hoc state checking
Provide a simpler graph-internal parsing and more detailed dependency
parsing assertions instead.

We had to temporarily pickle the graph in a fixed container-shared
directory before, only adding to the pollution the scan test (or
rather scan hack) does in general to achieve its goal. While this
commit only works for off/image states, the follow up commit will
enable this for on/vm states too and explain why we couldn't do this
before. However, we eagerly get rid of the scan test here to not
only rid ourselves of the brand new limitations but previous ones
where we had to do tricks like providing the graph as a static
attribute (was possible while the VT tests were run within the same
process and thus memory layout), adding awareness of which container
the scan test is running in and for which container it is checking
state availabily, and lots more signals that a single monolithic
state scanning is not supposed to happen within a single starting
test.

Did I even mention that the ad-hoc state checking aligns better with
our new notions of pool state backend as well as ad-hoc root state
provisioning?