Implement parallel lookahead runs on a fixed number of container slots
This immiates bash wrapped runs we have performed in the past and
handles the fact that currently we are not able to auto-setup LXC
containers (done with bash scripts externally).

Most importantly though we perform lookahead analysis during traversal
to identify pockets of parallelizable tests that we could traverse
at the same time.

Also improve the current traverse path logging.