Restate the on/off vm switch as subrequirements of the root states
In this way we can safely switch off vms as nonrunning vm is a
requirement for manipulating previously off states. At the same
time we can safely create missing images to boot a vm as a
requirement for manipulating previously on states.

We can also enforce these roots during state scanning in order
to detect otherwise detectable states in exchange of some additional
booting and image creating.