Add IP sniffing during vmnet default state retrieval
The vmnet state was moved to the off preprocessing hook to resolve
issues where the vm could be started before it was retrieved, thus
leading to non-existing bridges and other missing networking. The
problem is that this is not an ideal stage for it either and needs
at least some env processing networking steps. Since the sniffing
setup seems to be the only one of importance and it allows being
started earlier on, we prepare it earlier on.