Cache the already parsed nets or at least their parsing recipes
A node can thus be parsed starting from a net test object which
also determines all of its objects through state composition and
is analogous to the test net in terms of composition scope.

All nets are stored as final obtained objects as well as in the
intermediary graph in order to be reused among test leaves.