Fully support right-compatible variants in addition to left-compatible ones
This drops the previous default behavior where `only` filter from
a test node to a test object would apply just to the suffix specified
by `main_vm` and now makes explicit use of `only` as restrictive
for all joined suffices or suffixed as `only_<suffix>` to apply to
specific pre-joined variants.

In order to do this properly we also have to filter out excessively
large test nets that contain more than the few restricted vms that
a test net is reused for.