Convert all python unit tests to avocado unit tests
Besides replacing the test class (which will make the unit tests
detectable by Avocado and not just the test files which is the
default without the conversion) we also had to:

1) Remove any class setup methods as Avocado seems to ignore them
completely (and not just rerun them for each test process).

2) Perform more careful mock patching as Avocado seems to violate
the tested module's namespace calling the mock for its own test
running functionality.