Refactor the deploy/internal manual tools for vm/image state support
The new implementation drops state chains (with manual steps and
other chains to gradually be phased out later on) for the sake of
multi-type states support within the same invocation. It also drops
the previous "stateless" parameter for the sake of empty values for
the parameters of the updated states and a default stateful run for
the sake of a simple replacement with "setup=run only=all..deploy"
and a regular invocation through the run tool.