Formalize pool state backend compatibilty for collect/create/clean tools
This makes the clean step faster and makes the create step compatible
with the state pool backend to begin with. If we want to create or
clean vms, the expected intent is always for local vms.

A newly introduced tool for downloading a vm from the state pool is
called "collect" and is based on get root calls in comparison to
the set root for "create" and unset root for "clean".