Provide proper dynamic classes in the VMNetwork constructor
The vm network component types were not perfectly customizable and
needed some hacks like noop setters or simple enough classes in the
past in order to make this work. With the current cahnge we make
it easier for inheriting vm network classes to use their own
component types.

Thanks to Samir Aguiar for his final contribution to the plugin.