Provide decentralized pool for better IO distribution
We provide remote IP from a predefined nets range rather than
expecting /etc/hosts or other DNS configuration to resolve the
remote host's IP from its hostname for us.

The state scanning is still not properly localized but is where
we want to get to with this subbranch. A worker will leave a
"chemical scent" when it traverses a test node and any worker that
traverses the same node afterwards will scan for missing states
that are later meant to be local.