Introduce an occupation timeout to avoid infinite traversal durations
If a worker keeps waiting for the same occupied node to be free and
it never gets free (e.g. due to a crashed process of the worker that
occupied it), it has to drain its patience in accordance with an
anticipated test duration and log it on each attempt. Its sleep time
can also be determined from the test duration as a reasonably small
and now fully configurable value thus reducing some of the unit tests
duration along the way.