Rewrite the state subpackage unit tests for maximum driver agnosticity
We want to preserve the maximum coverage of actually executed code
(runtime validation) while reusing as much of the shared backend
logic as possible which leads us to a driver mocking context manager
and more-to-less-default assertion reuse. Remaining limitations of
both the state check mocking as well as the reusable assertions will
be addressed through development further implementations as well as
new higher level logic regarding root states.