Refactor pool state transfers into functions with remote support
Currently the implementation is limited to most basic aexpect use
but it could easily be extended to using aexpect's remote door to
invoke the remote state backends via a control file. This could
e.g. be the case for both a remote_backend.show() as well as a
remote_backend.unset() calls.