Get states implicitly for the state pool similar to roots pool
Even though backing chains of states are expected to be handled
with recursive downloading and the following pool state handling
is implicit rather than explicit, it is more compatible with the
way all pool state backends work at the moment and is still needed
in order for the state reuse to work with the decentralized pool
approach to follow next (i.e. set state based URL overwriting).