Scan locally for states on a container using a second control file
Second control file with respect to the first pre-test control file
which is also obligatory in order for us to run tests as subprocesses,
isolated or otherwise.

Since the remote host or container might be booting or in order
state of unreadiness, let's wait for the login and thus session
creation to succeed within an admissible timeout.