Extend the visited nodes to worker-specific visits
This solves special conditions when a worker would traverse a test
node some of whose setup has just been provided by other workers
but the test node itself has not yet been traversed by these workers.
The way we handle this situation is by making sure that a test node
is setup ready with respect to a specific slot and thus for a specific
worker and so is picking setup or cleanup nodes. A worker will thus
always make sure to traverse all setup including the reusable one that
just came fresh out of the oven.

This also completes the implementation for the test contract (isolated
test) established over the previous two commits on which workers should
provide what setup and run what tests but does so in less than ideal
fashion as the visited node logic is in great need of simplification.