Add support for more refined backing state chain downloads/uploads
In particular, trace the backing chain of images as external states
for a given image and download/upload those too in order to make sure
the state is actually downloaded/uploaded.

In case of any exceptions name sure to delete at least the initial
state so that further scanning could try downloading/uploading agin.