Create clean empty images on ramfile boot (set root) as an exception
Make sure to create clean empty images when passively keeping the
ones we already have does not work as expected. Always wiping these
will not combine with separately set image states (not as part of
the ramfile state) and will destroy the latter. Passively accepting
any images and just booting the vm could also result in the vm not
booting at all due to corrupted images (e.g. dirty layer with remoted
backing files). Thus, the current workaround is simply to try and boot
the vm and failing to do so (assumed to be caused by such corruption)
will then trigger wiping the images clean. The plan then is to simplify
this with future refactoring of the root state mechanism.

Note that we can try not to require booted vm when root checking and
boot only on set_root to make sure no vm is prematurely booted (when
root setting before state checking) before getting clean image states,
however this will result in the vm not booting at all since its root
(boot) state will be considered provided. This is clearly a problem of
all complexity that the root states have accumulated through years of
adding features with little to no redesigning of the logic.