Give more patience for obtaining the current test result
There are cases with missing test status simply due to the result
propagation taking some extra time. This is especially visible on
terminal nodes where not finding the results and decidng on ERROR
results in the vm not being installed and as a result cascading
failures down the dependency chain.