Allow container-unique names for test folders
This is the first step in allowing the same test node to be rerun
by multiple workers and as a result in multiple containers. In
particular, right now a worker would automatically retry a failed
or errored test with missing reusable setup as means to recreate
the setup for itself and all others to follow. Each test is thus
retried at least N times where N is the number of workers all with
a setup-governing policy.

While this does not exactly enable truly parallel retries, among
other things it also improves readability and leads to faster
debugging.