Separate image directories for external qcow2 snapshots
Similarly to the internal QCOW2 snapshots, we want to be able to
sometimes have two images with the same state name which would not
be possible if they are all placed in the same images directory.
This is especially important for vm state backends that would need
to match a vm state with both the states of memory and of each image.

In order to preserve compatibility with the previous image path
handling and since images could be considered as vm-level components
together with a potential vm memory ramfile, we keep these "pointer"
images within the vm (images) directory.