Save RAM by shutting down vms that are not needed by any test
This still leaves the possibility to leave vms running after a
test for debugging purposes and needs an Avocado VT pull request
like many other commits which should be merged by the time of release.

Later note: An Avocado VT version with this fix and all previous
fixes of great enough importance is 100. Unfortunately we don't
have sufficient control over the release process and resetting any
branches there to keep pointers to precise commit references. We
instead use the small use base for the Avocado I2N plugin as an
advantage here and simply make sure to have a synced version on LTS.
In the future we will improve on this further by providing exact
pointers to our own forks with middle versions.