Provide fixed mac address to host side bridges to deal with ARP cache
It would be also better test isolation if the vms cannot feel
that they exist in a different container (different remote hosts
that are supposed to be the same remote host) and this is one
way to ensure such consistency.

Also improve DNS debug info regarding vmnet host services in a
quick roundup master commit about small vmnet improvements instead
of a complete branch.