Extend test retrying to all available and multiple stop statuses
This behavior extends the capabilities of test node retrying to
multiple stop statuses and is more similar to otherwise scripted
tools we have already used before.

It also brings us one step closer to parallel retries in the future
since these should be based on analysis of the test results shared
across all workers and in particular their statuses.