Move from graph-scope set-based to test-scope get-based pool downloads
In order to make this possible we have to separate the locations
of getting/downloading and setting/uploading states both now using
the shared image pool as default only if not explicitly set. This
then also requires allowing state sync from the shared pool as well
which is a simpler overall sync behavior.

A worker will reset the corresponding location pointers only if it
actually ran or first visited the test and is therefore the newest
and most up to date provider of the corresponding test setup. This
is also simpler behavior than each worker modifying the pointers
upon traversal which could lead to circular resetting and other
potential problems we would like to avoid for future features.

Finally, a small simplification for a public chain transfer allows
for better reuse of parameters and more explicit API arguments.

Coupled with now simpler and explicit pool downloads (state/root
getting), this allows state downloads in parallel within each test
process and the traversal focusing on pure state scanning or checking
for any decision whether to run a given test and state cleaning or
syncing to guarantee identical setup in each worker environment.