Do not get states implicitly for the state and root pool
This means we now have a more standard behavior for the pool state
checking without any implicit state downloads to be done on the graph
traversal level. External state syncs will be performed by the graph
reversal procedures while actual state get downloads for tests must be
performed by within tests and thus in parallel.

This involves changes to the contracts for state and root checking
and getting with some simplifications while still having state/root
separation. An exception that prevents the completeness of these
simplifications is checking for the boot state via the sourced root
backend.