Verify external state hashes from state pool
In this way we make sure to have identical overlay images and as a
result identical backing chains. Downloading an external state that
is not compatible with the already cached "blockchain" of such will
result in invalid and thus corrupted image and vm states.

We hash only a chunk of the images since they can vary in size with
some vm images so large that a regular remote command would time out
(not to mention accumulated performnace penalty).