Do not rerun tests whose setup is lacking but cleaned
This small clause trades full compatibility with current logic and
minimal added code for even more complexity and should be removed
for the sake of more ground-breaking changes later on.

It also does not remedy the situation completely since while we will
not rerun a test simply because it was fully traversed and reversed
by another worker, we will still download reusable setup from other
sources having the current removable setup in their backing chain and
possibly not clean it up properly as e.g. only image states will be
downloaded along the chain and potential vm states will be incomplete.

The actual remedy and needed ground-breaking change is to consider node
reversal as an opportunity to either clean (f) reusable setup or sync it
(r) as an extra handling of reusable setup (now meaning not just keep
but also equalize across the worker swarm).