Move state sync functionality to each node similarly to state scans
In order to minimize usage of control files we reuse the pre-state
control file by extending it to multiple state actions.

Also improve the scan location verbosity in comparison.

In order to validate this properly we also have to make the dummy
localized state control more similar to the dummy state running, i.e.
more assertive as well as more encapsulated and more easily patched.
This then needs proper coverage for both cleaning and syncing states,
the first of which was available before but not properly tested.