Improve debug logging from logless isolated tests
Isolated tests are supposed to be small and isolated enough so that
they can pinpoint the cause of failure without any logging. However,
this is still needed in some cases to understand the tested unit's
behavior and rather indicates this behavior will have insufficient
logging in any general use case beyond the test itself.