Convert pool state backends to cross-environment extensions for others
This allows us to also better test in isolation and define appropriate
test contract for the pool functionality, moving it away from the
boundary class and its overly elaborate mocking.

The main motivation behind this change is that previous pool state
backends are unlike net state backends. The former act rather as cross-
environment extensions of regular state backends and should thus not
be composed of vm or lower backends. The latter is the actual composite
of vms and their images as backends and focuses on building an overall
network state, connecting and combining the vms it is composed of all
while staying within the environment and thus worker it resides in.

We also use this refactoring to simplify away the pool_only parameter
and leave todos for other excessively complex behaviors resulting from
all the years of small workaround extensions of the pool functionality.