Compare backing chains of states refreshing local cache if needed
This changes the implicit get behavior of both sourced (pool) state
backends whereby a cache state (if external) or cache root state (if
using internal states) will be refreshed if comparing it with the
one in the pool (and its complete backing chain) invalidates it. This
is crucial to detect overwritten dependencies higher in the backing
chain that would invalidate the currently cached external state in
the first case and could help refresh all internal states in the
second case.