Test equalized (synced) state setup across all swarm environments
Once a test completes we have to make sure all previously achieved,
shared, and reused setup across the worker swarm is available in a
way that is independent of which test set we would run next. In other
words, no worker should recreate the setup and thus mutate it if it
happens to traverse a path others have already traversed from a previous
run.