Exclude self-sync cases from the standard node reversal
Even though technically self-sync should result in a noop effect
this speeds up the reversal by not having to deploy control files
and execute them for the last worker that has originally traversed
the test node (and with later functionality every worker that has
run the test). It will also not be better handled with future pool
get extensions that could use multiple pool sources since a worker
is its *own* source of the synced setup.

The main side effects of this are the refactoring of node's session
and IP environment methods which marks the beginning of the branch
generalization to distributed testing.