Restore backwards compatibility with process spawner serial runs
This allows us to use supercontainers or hosts composed of containers
as a more basic setup where the tests will not be isolated and run
in a serial manner as processes instead.

All we needed for this is to allow reflexive (self) ssh login into
localhost as a way to locally scan for external states and assume
a process spawner if no slots for subcontainers were provided.