Separate read-only state data within the external state backends
This can be used to easily base the entire swarm of workers on top
of the same data storage and make them push to the swarm pool as
state setting. The immediate advantage is using a lot less storage
space as each worker does not keep a separate copy and there is only
one read-only state copy per swarm.

It is possible to use only when there is an explicit knowledge that
the worker swarm is not based over an entire cluster and distributed
across different hosts which is always the case at present (using LXC
spawner only and thus only containers to isolate the tests). Before
any possibility of distributed testing is possible though we have to
simplify various aspects of the graph traversal to make the parallel
traversal even more natural in the current code base.

This is also a much easier alternative to the symlink setup since it
has symmetric relations without one copy of the setup being a mere
(asymmetrically related) mirror of the original.

Finally, to test this restore the qcow2ext isolation test coverage
which was abandoned as a "boundary needing integration tests" but
until such tests are available we have to test in other ways. Not
to mention that this change is going against the imperative boundary
and thus at the very least needs necessary contract validation for
the structured programming logic.