Remove aexpect and avocado-vt from the RPM build dependencies
The build process only makes use of avocado to run the isolation
tests. The above dependencies were added before because of a
circular complexity in avocado's extension manager that would
end up not just running the isolation tests but also loading the
Avocado I2N plugin and as a result require install dependencies
in addition to RPM build dependencies. While this is still not
fixed at present, the above would at least work well for cases
where we build the RPM but do not run isolation tests and reduce
the complexity at least to an extend.