Move all nonpermanent vms to qcow2ext/ramfile state backends
The qcow2/qcow2vt combo is gradually moving to the attic just like
the oldest LVM image backend but the main reason for this change is
that it is not optimal for use in a swarm setup reuse where only
the first few internal snapshots will be downloaded together as a
bundle from the rest of the workers and each one would then rerun
the test adding its own further internal snapshots. A better way
to both use and test internal snapshots is with serial runs and
types of runs that go beyond our current integration tests.

The qcow2/qcow2vt combo is still used for the permanent vm vm3 as
most internal snapshots could be provided in advance and it is
easier to move permanent vms around as one bundle of states but
this is also to still provide minimal coverage for this combo.

The most likely way for a future support of these backends will
be similar to the LVM - we keep the code for anyone that might
still have interest in reviving/maintaining it but the feature
would most likely gradually degrade and break down. There are
simply not enough many resources to support these backends with
maximum effort and the new methods we employ are clearly superior
for the parallel run advancements.