Add initial support for multiple pool locations and scopes
The change in traversal now allows for registering each worker as
an additional provider of a reusable setup if it ran the test and
passed and not if it skipped the test, had a failure or error as
a test status, or if the test was the shared root. The most work
is then in extending the pool state backends with the capability
to use multiple locations of multiple scopes for transport.

A major simplification in this is making check/get/set/unset
locations obligatory with the shared pool being explicitly
provided during traversal as a default extra location.

This includes additional isolation tests of the new concepts here
like correct scope blocking, source parameter propagation, etc.

The root source state backends will be reimplemented and fully
reconsidered next.