Take advantage of vm-object independent ramfile state unset
This advantage is generally used via a "use_env=no" parameter to
perform some vm state operations outside of tests that provide the
complete vm environment and in this case is needed for the clean up
control that is run during node reversal.

Also increase the sample test suite diversity to cover the case of
unset vm states in addition to unset image states.