Consider multiple swarms to decide whether to run temporary nodes
The old todo we tried to get rid of in this branch will in fact
survive it and has to be extended to support correct run decisions
for multiple possibly filtered swarms where each worker should make
the previous evaluation based on swarm-bound information.

This is hopefully going to be treated further together with the
current limitation on filtered swarms queuing up on separately
provided setup by generalizing the `is_occupied` functionality
to multiple past and present workers of different swarms allowing
for node re-entrancy under sufficient safety conditions.