Deploy avocado dependencies via cloning for CI isolation tests
The main reason not to use PyPI packages is that Avocado VT's
deployment pipeline there has been discontinued. Note that our
pip dependencies file is not wrong in specifying the version
for Avocado VT since if it is already installed the dependency
will be satisfied and we cannot control and don't specify the
way in which Avocado VT should be deployed to begin with.

This approach is also easier for testing with modified versions
of the avocado dependencies since it doesn't require the complete
package publishing process and additional stages.