Restore test output directories for Avocado versions 102+
This adapts to new behavior in avocado and our merged LXC spawner
that was introduced after 101 and before 102. In particular, out
custom scheduler must take care of the test output directory setup
and Avocado VT doesn't need this because it uses the default avocado
scheduler.

Take the opportunity to also imitate the full name URI setup for
each test like in Avocado VT and avoid cases of folder collision
due to identical shorter test names.