Recreate dead sessions from tests without a retrieved status
The optional remote spawner use for distributed testing is not yet
fully implemented, reviewed, and polished and its use of its aexpect
dependency could lead to various hiccups with hanging or dead sessions
that we have to gradually get a better hold of with more progress on
the spawner. Therefore as an extra precaution and for the time being
let's make our scheduler's session and thus remote spawner management
more robust to such problems in order to avoid cascading problems
over the entire test run.

So far the new session health check empirically caught and thus
reported and circumvented a few of the verified problems but these
remain to be ironed out on a deeper level once upstream catches up
with our pull requests.